Move some things for clarity
